requirements.txt
fastapi==0.88.0
pytest==7.2.0
SQLAlchemy==1.4.46
uvicorn==0.20.0
httpx==0.23.3
PyJWT==2.6.0